requirements/dev.txt
fastapi[all]==0.110.2
sqlalchemy==2.0.29
fastapi-users[sqlalchemy]==13.0.0
asyncpg==0.29.0
psycopg2==2.9.9
uvicorn==0.29.0
python-dotenv==1.0.1
alembic==1.13.1
fastapi-cache2[redis]==0.2.1
flower==2.0.1
pytest==8.1.1
pytest-asyncio==0.23.6